Add files via upload
Use this origin_report as a Final Traveler Report.  Upload it into the app to perform the test. Ensure 'Run Model C Detection is clicked on the left sidebar. The 3 check boxes beneath that should be checked as well.  In the main area, under 'Select report Time & Date', select 'Choose a time'.  Use 2025/06/26 as the date and 08:00 as the time. Under 'Select Day Start Time, select 18:00.  Under Scope, select 'Days'.  The B model and C Model Examples file shows the models that are being missed.